Fix Reviews endpoins

+++ src/api/reviews/index.js
@@ -3,27 +3,49 @@
 import express from "express"
 import createHttpError from "http-errors"
 import ReviewModel from "./model.js"
+import ProductModel from "../products/model.js"
 
 const reviewRouter = express.Router()
 
-reviewRouter.post("/", async (req, res, next) => {
+reviewRouter.post("/:productId/reviews", async (req, res, next) => {
   try {
     const newReview = new ReviewModel(req.body)
     const { _id } = await newReview.save()
+    const findProduct = await ProductModel.findByIdAndUpdate(
+      req.params.productId,
+      { $push: { reviews: _id } },
+      { new: true, runValidators: true }
+    )
+    if (!findProduct)
+      return next(
+        createHttpError(
+          404,
+          `product with id: ${req.params.productId} not foud`
+        )
+      )
     res.status(201).send({ _id })
   } catch (error) {
     next(error)
   }
 })
-reviewRouter.get("/", async (req, res, next) => {
+reviewRouter.get("/:productId/reviews", async (req, res, next) => {
   try {
-    const reviews = await ReviewModel.find()
+    const reviews = await ProductModel.findById(req.params.productId).populate(
+      "reviews"
+    )
+    if (!reviews)
+      return next(
+        createHttpError(
+          404,
+          `product with id: ${req.params.productId} not foud`
+        )
+      )
     res.send(reviews)
   } catch (error) {
     next(error)
   }
 })
-reviewRouter.get("/:reviewId", async (req, res, next) => {
+reviewRouter.get("/reviews/:reviewId", async (req, res, next) => {
   try {
     const review = await ReviewModel.findById(req.params.reviewId)
     if (review) {
@@ -35,7 +57,7 @@ reviewRouter.get("/:reviewId", async (req, res, next) => {
     next(error)
   }
 })
-reviewRouter.put("/:reviewId", async (req, res, next) => {
+reviewRouter.put("/reviews/:reviewId", async (req, res, next) => {
   try {
     const updatedReview = await ReviewModel.findByIdAndUpdate(
       req.params.reviewId,
@@ -51,16 +73,31 @@ reviewRouter.put("/:reviewId", async (req, res, next) => {
     next(error)
   }
 })
-reviewRouter.delete("/:reviewId", async (req, res, next) => {
+reviewRouter.delete("/:productId/reviews/:reviewId", async (req, res, next) => {
   try {
-    const deleteReview = await ReviewModel.findByIdAndDelete(
+    const updatedProduct = await ProductModel.findByIdAndUpdate(
+      req.params.productId,
+      { $pull: { reviews: { _id: req.params.reviewId } } },
+      { new: true, runValidators: true }
+    )
+    if (!updatedProduct)
+      return next(
+        createHttpError(
+          404,
+          `product with id: ${req.params.productId} not foud`
+        )
+      )
+
+    const deletedReview = await ReviewModel.findByIdAndDelete(
       req.params.reviewId
     )
-    if (deleteReview) {
-      res.status(204).send()
-    } else {
-      next(createError(404, `Review with id ${req.params.reviewId} not found!`))
-    }
+
+    if (!deletedReview)
+      return next(
+        createHttpError(404, `review with id: ${req.params.reviewId} not foud`)
+      )
+
+    res.status(204).send()
   } catch (error) {
     next(error)
   }
